+++ Assets/_Scripts/Explorer/Systems/CombatSystem/GameplayEffectSpec.cs
@@ -7,7 +7,7 @@ namespace Explorer._Scripts.Explorer.Systems.CombatSystem
     public class GameplayEffectSpec
     {
         public GameplayEffectDef Def;
-        public float Level;
+        public int Level;
         public GameObject Instigator;
         public Dictionary<string, float> ResolvedMagnitudes = new(StringComparer.OrdinalIgnoreCase);
     }
